Move bids into separate storage

+++ src/interfaces/IAuction.sol
@@ -28,11 +28,11 @@ interface IAuction is ITickStorage, IAuctionStepStorage {
     error FundsRecipientIsZero();
 
     /// @notice Emitted when a bid is submitted
-    /// @param id The id of the tick
+    /// @param id The id of the bid
     /// @param price The price of the bid
     /// @param exactIn Whether the bid is exact in
     /// @param amount The amount of the bid
-    event BidSubmitted(uint128 indexed id, uint256 price, bool exactIn, uint256 amount);
+    event BidSubmitted(uint256 indexed id, uint256 price, bool exactIn, uint256 amount);
 
     /// @notice Emitted when a new checkpoint is created
     /// @param blockNumber The block number of the checkpoint
